+++ src/graphql/schema.js
@@ -87,6 +87,7 @@ type MultipleResultsEvent implements Transaction {
   block: Block
   address: String
   ownerAddress: String!
+  ownerName: String
   version: Int
   name: String!
   results: [String!]!
@@ -133,6 +134,7 @@ type Bet implements Transaction {
   block: Block
   eventAddress: String!
   betterAddress: String!
+  betterName: String
   resultIndex: Int!
   amount: String!
   eventRound: Int!
@@ -155,6 +157,7 @@ type ResultSet implements Transaction {
   block: Block
   eventAddress: String!
   centralizedOracleAddress: String
+  centralizedOracleName: String
   resultIndex: Int!
   amount: String!
   eventRound: Int!
@@ -179,6 +182,7 @@ type Withdraw implements Transaction {
   block: Block
   eventAddress: String!
   winnerAddress: String!
+  winnerName: String
   winningAmount: String!
   escrowWithdrawAmount: String!
   eventName: String
@@ -218,6 +222,7 @@ type AllStats {
 type MostBet {
   eventAddress: String
   betterAddress: String!
+  betterName: String
   amount: String!
 }
 
@@ -230,6 +235,7 @@ type PaginatedMostBets {
 type BiggestWinner {
   eventAddress: String!
   betterAddress: String!
+  betterName: String
   amount: String!
 }
 
@@ -242,6 +248,7 @@ type PaginatedBiggestWinner {
 type LeaderboardEntry {
   eventAddress: String
   userAddress: String!
+  userName: String
   investments: String!
   winnings: String!
   returnRatio: Float!