+++ src/sync/bet-placed.js
@@ -1,7 +1,7 @@
 const { each, isNull, find } = require('lodash');
 const web3 = require('../web3');
 const { TX_STATUS } = require('../constants');
-const { toLowerCase } = require('../utils');
+const { toLowerCase, getAndInsertNames } = require('../utils');
 const logger = require('../utils/logger');
 const { getTransactionReceipt } = require('../utils/web3-utils');
 const DBHelper = require('../db/db-helper');
@@ -27,7 +27,7 @@ const syncBetPlaced = async ({ startBlock, endBlock, syncPromises, limit }) => {
           // Parse and insert bet
           const bet = parseBet({ log: logObj });
           await DBHelper.insertBet(bet);
-
+          await getAndInsertNames(bet.betterAddress, DBHelper);
           // Fetch and insert tx receipt
           const txReceipt = await getTransactionReceipt(bet.txid);
           await DBHelper.insertTransactionReceipt(txReceipt);